20231128 @Mookse - closes #98 - factory only required for dynamic classes when extender requires them
